Development

+++ app/Domains/Image/Services/ImageService.php
@@ -37,10 +37,13 @@ class ImageService extends BaseService
         $name = str_replace(' ', '_', strtolower($data['name']));
 
         try {
+
             $this->model->create([
                 'name' => $name,
             ]);
+
             Storage::disk('public')->makeDirectory("carousel/$name");
+
         }catch (\Exception $e) {
             DB::rollBack();
             throw new GeneralException('There was a problem creating the image.');
@@ -51,27 +54,42 @@ class ImageService extends BaseService
     }
 
     /**
+     * Upload image
+     * @param array $data
+     */
+    public function upload(array $data = [], string $id)
+    {
+        $imageInstance = $this->model->find($id);
+        foreach($data['image'] as $image) {
+            $image->storeAs("carousel/$imageInstance->name", $image->getClientOriginalName(), 'public');
+        }
+    }
+
+    /**
      * 
      */
     public function update(array $data, string $id)
     {
-        $image = Image::find($id);
+        $image = $this->model->find($id);
         DB::beginTransaction();
 
         try{
 
             // Update Directory name
-            $files = Storage::disk('public')->allFiles("carousel/$image->name");
-            Storage::disk('public')->makeDirectory("carousel/" . $data['name']);
-            foreach ($files as $file) {
-                $name = basename($file);
-                Storage::disk('public')->move("carousel/$image->name/$name", "carousel/" . $data['name'] . "/$name");
+            if(isset($data['name'])){
+                $files = Storage::disk('public')->allFiles("carousel/$image->name");
+                Storage::disk('public')->makeDirectory("carousel/" . $data['name']);
+                foreach ($files as $file) {
+                    $name = basename($file);
+                    Storage::disk('public')->move("carousel/$image->name/$name", "carousel/" . $data['name'] . "/$name");
+                }
+                Storage::disk('public')->deleteDirectory("carousel/$image->name");
             }
-            Storage::disk('public')->deleteDirectory("carousel/$image->name");
 
             // Update image name
             $image->update([
-                'name' => $data['name']
+                'name' => $data['name'] ?? $image->name,
+                'allow_roles' => $data['roles'] ?? $image->allow_roles,
             ]);
 
 
@@ -81,6 +99,23 @@ class ImageService extends BaseService
         }
 
         DB::commit();
+    }
+
+    /**
+     * Delete Image
+     * @param array $data
+     * @param string $id
+     */
+    public function delete(array $data, string $id)
+    {
+        $image = $this->model->find($id);
+        $imageFiles = Storage::disk('public')->allFiles("carousel/$image->name");
+        $index = array_search($data['name'], array_map('basename', $imageFiles));
+
+        if($index !== false) {
+            $filePath = $imageFiles[$index];
+            Storage::disk('public')->delete($filePath);
+        }
     }
 
     /**
